Extract chat input handling into UI::Chat
Move CheckChatText (routes a typed line to a command, whisper or message,
~165 lines) out of ZzzInterface.cpp into UI/Chat/ChatInput.{h,cpp} under
namespace UI::Chat. Its two callers are now qualified. No behaviour change.
