+++ src/source/Engine/Object/ZzzInterface.h
@@ -64,7 +64,6 @@ void MoveHero();
 void ClearInput(BOOL bClearWhisperTarget = TRUE);
 void CutText(const wchar_t* Text, wchar_t* Text1, wchar_t* Text2, size_t MaxLength);
 void RenderSwichState();
-void CheckChatText(wchar_t* Text);
 
 
 //  HotKey
